feat(CEPA-292): Create payment for teacher

+++ src/main/java/com/example/courseservice/data/entities/TeacherIncome.java
@@ -34,9 +34,9 @@ public class TeacherIncome {
     @GeneratedValue(strategy = GenerationType.AUTO, generator = "teacher_income_sequence")
     private long id;
 
-    private Long money;
+    private Double money;
 
-    private Long receivedMoney;
+    private Double receivedMoney;
 
     private Integer month;
